+++ __tests__/src/packages/imperative/plugins/suites/ListPlugins.ts
@@ -10,7 +10,7 @@
 */
 
 import * as T from "../../../../../src/TestUtil";
-import { cliBin } from "../PluginManagementFacility.spec";
+import { cliBin } from "../../__integration__/PluginManagementFacility.spec";
 import { join } from "path";
 
 describe("List plugin", () => {